+++ lib/stamina-core/stamina/utils/decorate.rb
@@ -1,24 +1,34 @@
 module Stamina
   module Utils
     #
-    # Decorates states of an automaton by applying a propagation rule
-    # until a fix point is reached.
+    # Decorates states of an automaton by applying a propagation rule until a fix point is
+    # reached.
     #
     class Decorate
 
-      # The key to use to maintain the decoration on states (:invariant
-      # is used by default)
-      attr_writer :decoration_key
-
       # Creates a decoration algorithm instance
-      def initialize(decoration_key = :invariant)
-        @decoration_key = decoration_key
+      def initialize(output = :invariant)
+        @output    = Decorate.state_output(output) if output
         @suppremum = nil
         @propagate = nil
         @initiator = nil
+        @backward  = false
         @start_predicate = nil
-        @backward = false
       end
+
+      # Builds an output hash that keeps decoration in states
+      def self.state_output(decoration_key)
+        Object.new.extend Module.new{
+          define_method :[] do |state|
+            state[decoration_key]
+          end
+          define_method :[]= do |state,deco|
+            state[decoration_key] = deco
+          end
+        }
+      end
+
+      ### CONFIGURATION ##################################################################
 
       # Installs a suppremum function through a block.
       def set_suppremum(&block)
@@ -32,16 +42,6 @@ module Stamina
         set_suppremum(&proc)
       end
 
-      # Computes the suppremum between two decorations. By default, this method
-      # looks for a suppremum function installed with set_suppremum. If not found,
-      # it tries calling a suppremum method on d0. If not found it raises an error.
-      # This method may be overriden.
-      def suppremum(d0, d1)
-        return @suppremum.call(d0, d1) if @suppremum
-        return d0.suppremum(d1) if d0.respond_to?(:suppremum)
-        raise "No suppremum function installed or implemented by decorations"
-      end
-
       # Installs a propagate function through a block.
       def set_propagate(&block)
         raise ArgumentError, 'Propagate expected through a block' if block.nil?
@@ -52,16 +52,6 @@ module Stamina
       # Same as #set_propagate, but with an explicit proc
       def propagate=(proc)
         set_propagate(&proc)
-      end
-
-      # Computes the propagation rule. By default, this method looks for a propagate
-      # function installed with set_propagate. If not found, it tries calling a +
-      # method on deco. If not found it raises an error.
-      # This method may be overriden.
-      def propagate(deco, edge)
-        return @propagate.call(deco, edge) if @propagate
-        return deco.+(edge) if deco.respond_to?(:+)
-        raise "No propagate function installed or implemented by decorations"
       end
 
       # Set an initiator methods, responsible of computing the initial decoration of
@@ -77,12 +67,6 @@ module Stamina
         set_initiator(&proc)
       end
 
-      # Returns the initial decoration of state `s`
-      def init_deco(s)
-        return @initiator.call(s) if @initiator
-        raise "No initiator function installed"
-      end
-
       # Sets the start predicate to use
       def set_start_predicate(&block)
         raise ArgumentError, 'Start predicate expected through a block' if block.nil?
@@ -95,15 +79,43 @@ module Stamina
         set_start_predicate(&proc)
       end
 
+      # Sets if the algorithms works backward
+      def backward=(val)
+        @backward = val
+      end
+
+      ### SUBCLASS HOOKS #################################################################
+
+      # Computes the suppremum between two decorations. By default, this method
+      # looks for a suppremum function installed with set_suppremum. If not found,
+      # it tries calling a suppremum method on d0. If not found it raises an error.
+      # This method may be overriden.
+      def suppremum(d0, d1)
+        return @suppremum.call(d0, d1) if @suppremum
+        return d0.suppremum(d1) if d0.respond_to?(:suppremum)
+        raise "No suppremum function installed or implemented by decorations"
+      end
+
+      # Computes the propagation rule. By default, this method looks for a propagate
+      # function installed with set_propagate. If not found, it tries calling a +
+      # method on deco. If not found it raises an error.
+      # This method may be overriden.
+      def propagate(deco, edge)
+        return @propagate.call(deco, edge) if @propagate
+        return deco.+(edge) if deco.respond_to?(:+)
+        raise "No propagate function installed or implemented by decorations"
+      end
+
+      # Returns the initial decoration of state `s`
+      def init_deco(s)
+        return @initiator.call(s) if @initiator
+        raise "No initiator function installed"
+      end
+
       # Returns the start predicate
       def take_at_start?(s)
         return @start_predicate.call(s) if @start_predicate
         raise "No start predicate function installed"
-      end
-
-      # Sets if the algorithms works backward
-      def backward=(val)
-        @backward = val
       end
 
       # Work backward?
@@ -111,32 +123,55 @@ module Stamina
         @backward
       end
 
+      ### MAIN ###########################################################################
+
       # Executes the propagation algorithm on a given automaton.
-      def call(fa)
-        fa.states.each do |s|
-          s[@decoration_key] = init_deco(s)
-        end
-        to_explore = fa.states.select{|s| take_at_start?(s)}
-        until to_explore.empty?
-          source = to_explore.pop
-          (backward? ? source.in_edges : source.out_edges).each do |edge|
-            target = backward? ? edge.source : edge.target
-            p_decor = propagate(source[@decoration_key], edge)
-            p_decor = suppremum(target[@decoration_key], p_decor)
-            unless p_decor == target[@decoration_key]
-              target[@decoration_key] = p_decor
-              to_explore << target unless to_explore.include?(target)
+      def call(fa, out = nil)
+        with_output(out) do |output|
+          fa.states.each{|s| output[s] = init_deco(s) }        # Init decoration on each state
+          to_explore = fa.states.select{|s| take_at_start?(s)} # Init to_explore (start predicate)
+          until to_explore.empty?                              # empty to_explore now
+            source = to_explore.pop
+            each_edge_and_target(source) do |edge, target|
+              p_decor = propagate(output[source], edge)
+              p_decor = suppremum(output[target], p_decor)
+              unless p_decor == output[target]
+                output[target] = p_decor
+                to_explore << target unless to_explore.include?(target)
+              end
             end
           end
+          fa
         end
-        fa
       end
 
       # Executes the propagation algorithm on a given automaton.
       def execute(fa, bottom, d0)
+        warn "Decorate#execute is deprecated, use Decorate#call (#{caller[0]})"
         self.initiator       = lambda{|s| (s.initial? ? d0 : bottom)}
         self.start_predicate = lambda{|s| s.initial? }
         call(fa)
+      end
+
+    private
+
+      def with_output(out)
+        if out
+          yield out.is_a?(Symbol) ? Decorate.state_output(out) : out
+        elsif @output
+          warn "Decorate.new(:decokey) is deprecated, use Decorate#call(fa, :decokey) (#{caller[1]})"
+          yield @output
+        else
+          raise ArgumentError, "Output may not be nil", caller
+        end
+      end
+
+      def each_edge_and_target(source)
+        edges = backward? ? source.in_edges : source.out_edges
+        edges.each do |edge|
+          target = target = backward? ? edge.source : edge.target
+          yield edge, target
+        end
       end
 
     end # class Decorate