+++ test/unit/utils/decorate_test.rb
@@ -4,16 +4,22 @@ module Stamina
     class DecorateTest < ::Stamina::StaminaTest
 
       module Reachability
+        def init_deco(s) s.initial?; end
+        def take_at_start?(s) s.initial?; end
         def suppremum(d0, d1) d0 || d1; end
         def propagate(deco, edge) deco; end
       end
 
       module Depth
+        def init_deco(s) s.initial? ? 0 : 1000000; end
+        def take_at_start?(s) s.initial?; end
         def suppremum(d0, d1) (d0 < d1 ? d0 : d1) end
         def propagate(deco, edge) deco+1; end
       end
 
       module ShortPrefix
+        def init_deco(s) s.initial? ? [] : nil; end
+        def take_at_start?(s) s.initial?; end
         def suppremum(d0, d1)
           return d0 if d1.nil?
           return d1 if d0.nil?
@@ -25,32 +31,34 @@ module Stamina
       end
 
       def test_reachability_on_small_dfa
-        algo = Stamina::Utils::Decorate.new(:reachable)
+        algo = Stamina::Utils::Decorate.new
         algo.set_suppremum {|d0,d1|  d0 || d1 }
         algo.set_propagate {|deco,edge| deco }
-        algo.execute(@small_dfa, false, true)
-        assert_equal @small_dfa.states.select {|s| s[:reachable]==true}, @small_dfa.states
+        algo.set_initiator {|s| s.initial?}
+        algo.set_start_predicate{|s| s.initial?}
+        algo.call(@small_dfa, :reachable)
+        assert_equal @small_dfa.states.select{|s| s[:reachable]==true}, @small_dfa.states
 
-        algo = Stamina::Utils::Decorate.new(:reachable)
+        algo = Stamina::Utils::Decorate.new
         algo.extend(Reachability)
-        algo.execute(@small_dfa, false, true)
-        assert_equal @small_dfa.states.select {|s| s[:reachable]==true}, @small_dfa.states
+        algo.call(@small_dfa, :reachable)
+        assert_equal @small_dfa.states.select{|s| s[:reachable]==true}, @small_dfa.states
       end
 
       def test_depth_on_small_dfa
-        algo = Stamina::Utils::Decorate.new(:depth)
+        algo = Stamina::Utils::Decorate.new
         algo.extend(Depth)
-        algo.execute(@small_dfa, 1000000, 0)
+        algo.call(@small_dfa, :depth)
         assert_equal 0, @small_dfa.ith_state(3)[:depth]
         assert_equal 1, @small_dfa.ith_state(2)[:depth]
         assert_equal 2, @small_dfa.ith_state(0)[:depth]
         assert_equal 3, @small_dfa.ith_state(1)[:depth]
       end
 
-      def test_depth_on_small_dfa
-        algo = Stamina::Utils::Decorate.new(:short_prefix)
+      def test_short_prefix_on_small_dfa
+        algo = Stamina::Utils::Decorate.new
         algo.extend(ShortPrefix)
-        algo.execute(@small_dfa, nil, [])
+        algo.call(@small_dfa, :short_prefix)
         assert_equal [], @small_dfa.ith_state(3)[:short_prefix]
         assert_equal ['b'], @small_dfa.ith_state(2)[:short_prefix]
         assert_equal ['b', 'c'], @small_dfa.ith_state(0)[:short_prefix]