+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# 0.5.5 / FIX ME
+# 0.6.0 / FIX ME
 
 * Generalized the decoration algorithm to work backwards as well as forwards as well as
   accepting an output Hash-like object, defaulting to true state decorations.
+* Generalized the DFA equivalence algorithm under TransitionSystem::Equivalence.
 
 # O.5.4 / 2012-03-06